# Improve  `chol_lower` and `chol_upper`