--- a/src/chol.jl
+++ b/src/chol.jl
@@ -1,7 +1,14 @@
 # Accessing a.L directly might involve an extra copy();
-# instead, always use the stored Cholesky factor:
-chol_lower(a::Cholesky) = a.uplo === 'L' ? a.L : a.U'
-chol_upper(a::Cholesky) = a.uplo === 'U' ? a.U : a.L'
+# instead, always use the stored Cholesky factor
+# Using `a.factors` instead of `a.L` or `a.U` avoids one
+# additional `LowerTriangular` or `UpperTriangular` wrapper and
+# leads to better performance
+function chol_lower(a::Cholesky)
+    return a.uplo === 'L' ? LowerTriangular(a.factors) : LowerTriangular(a.factors')
+end
+function chol_upper(a::Cholesky)
+    return a.uplo === 'U' ? UpperTriangular(a.factors) : UpperTriangular(a.factors')
+end
 
 # For a dense Matrix, the following allows us to avoid the Adjoint wrapper:
 chol_lower(a::Matrix) = cholesky(Symmetric(a, :L)).L