This PR improves `chol_lower` and `chol_upper` by using the field `factors` directly:
```julia
julia> using PDMats, LinearAlgebra, BenchmarkTools

julia> C = cholesky(Symmetric(Matrix(I, 250, 250), :L));

julia> @btime PDMats.chol_lower($C);
  # master: 11.995 ns (1 allocation: 16 bytes)
  # this PR: 8.097 ns (1 allocation: 16 bytes)

julia> @btime PDMats.chol_upper($C);
  # master: 15.072 ns (2 allocations: 32 bytes)
  # this PR: 8.106 ns (1 allocation: 16 bytes)

julia> C = cholesky(Symmetric(Matrix(I, 250, 250), :U));

julia> @btime PDMats.chol_lower($C);
  # master: 15.248 ns (2 allocations: 32 bytes)
  # this PR: 8.116 ns (1 allocation: 16 bytes)

julia> @btime PDMats.chol_upper($C);
  # master: 12.136 ns (1 allocation: 16 bytes)
  # this PR: 8.100 ns (1 allocation: 16 bytes)
```