@andreasnoack 2022-01-18
Is the type instability a concern?