@devmotion 2022-01-18
If you mean if it caused me problems in any downstream application, then no, it didn't. I noticed it when I played around with https://github.com/FluxML/Zygote.jl/pull/1117 and this PR seemed like a simple improvement. But I'm not too invested in the PR and therefore didn't push it, I just remembered it when I went through my list of open PRs yesterday.