@andreasnoack 2022-01-19
I was considering the type instability in the new version `a.uplo === 'U' ? UpperTriangular(a.factors) : UpperTriangular(a.factors')`. The alternative would be something like `a.uplo === 'U' ? UpperTriangular(a.factors) : UpperTriangular(copy(a.factors'))`. It would make it type stable at the cost of the copy.