@devmotion 2022-01-20
Ah, sorry, I misunderstood. The PR doesn't change this type instability, the return types are exactly the same as on the master branch. For completeness, I benchmarked the version on the master branch, the version in this PR, and the version in this PR + `copy`:
```julia
julia> using BenchmarkTools, LinearAlgebra

julia> chol_lower(a::Cholesky) = a.uplo === 'L' ? a.L : a.U'
chol_lower (generic function with 1 method)

julia> chol_upper(a::Cholesky) = a.uplo === 'U' ? a.U : a.L'
chol_upper (generic function with 1 method)

julia> function chol_lower_pr(a::Cholesky)
           return a.uplo === 'L' ? LowerTriangular(a.factors) : LowerTriangular(copy(a.factors'))
       end
chol_lower_pr (generic function with 1 method)

julia> function chol_lower_pr(a::Cholesky)
           return a.uplo === 'L' ? LowerTriangular(a.factors) : LowerTriangular(a.factors')
       end
chol_lower_pr (generic function with 1 method)

julia> function chol_lower_pr_copy(a::Cholesky)
           return a.uplo === 'L' ? LowerTriangular(a.factors) : LowerTriangular(copy(a.factors'))
       end
chol_lower_pr_copy (generic function with 1 method)

julia> function chol_upper_pr_copy(a::Cholesky)
           return a.uplo === 'U' ? UpperTriangular(a.factors) : UpperTriangular(copy(a.factors'))
       end
chol_upper_pr_copy (generic function with 1 method)

julia> C = cholesky(Symmetric(Matrix(I, 250, 250), :L));

julia> typeof(chol_lower(C))
LowerTriangular{Float64, Matrix{Float64}}

julia> typeof(chol_lower_pr(C))
LowerTriangular{Float64, Matrix{Float64}}

julia> typeof(chol_lower_pr_copy(C))
LowerTriangular{Float64, Matrix{Float64}}

julia> @btime chol_lower($C);
  12.179 ns (1 allocation: 16 bytes)

julia> @btime chol_lower_pr($C);
  8.839 ns (1 allocation: 16 bytes)

julia> @btime chol_lower_pr_copy($C);
  4.598 ns (0 allocations: 0 bytes)

julia> typeof(chol_upper(C))
UpperTriangular{Float64, Adjoint{Float64, Matrix{Float64}}}

julia> typeof(chol_upper_pr(C))
UpperTriangular{Float64, Adjoint{Float64, Matrix{Float64}}}

julia> typeof(chol_upper_pr_copy(C))
UpperTriangular{Float64, Matrix{Float64}}

julia> @btime chol_upper($C);
  16.554 ns (2 allocations: 32 bytes)

julia> @btime chol_upper_pr($C);
  8.933 ns (1 allocation: 16 bytes)

julia> @btime chol_upper_pr_copy($C);
  69.869 μs (2 allocations: 488.36 KiB)

julia> C = cholesky(Symmetric(Matrix(I, 250, 250), :U));

julia> typeof(chol_lower(C))
LowerTriangular{Float64, Adjoint{Float64, Matrix{Float64}}}

julia> typeof(chol_lower_pr(C))
LowerTriangular{Float64, Adjoint{Float64, Matrix{Float64}}}

julia> typeof(chol_lower_pr_copy(C))
LowerTriangular{Float64, Matrix{Float64}}

julia> @btime chol_lower($C);
  16.602 ns (2 allocations: 32 bytes)

julia> @btime chol_lower_pr($C);
  8.655 ns (1 allocation: 16 bytes)

julia> @btime chol_lower_pr_copy($C);
  69.314 μs (2 allocations: 488.36 KiB)

julia> typeof(chol_upper(C))
UpperTriangular{Float64, Matrix{Float64}}

julia> typeof(chol_upper_pr(C))
UpperTriangular{Float64, Matrix{Float64}}

julia> typeof(chol_upper_pr_copy(C))
UpperTriangular{Float64, Matrix{Float64}}

julia> @btime chol_upper($C);
  12.153 ns (1 allocation: 16 bytes)

julia> @btime chol_upper_pr($C);
  8.617 ns (1 allocation: 16 bytes)

julia> @btime chol_upper_pr_copy($C);
  4.588 ns (0 allocations: 0 bytes)
```

As expected the version with `copy` fixes the type instability, and thus gets rid of the remaining allocation if the factor is not copied. This leads also to better performance (for the example above 4.5ns vs 9ns with this PR). However, if the factors have to be copied performances is much much worse: in the example here it increases from around 9ns with this PR to around 70μs (!).

So I don't think it's worth fixing the type instability.